requirements.txt
Django==1.9.8
basescript==0.2.6
django-reversion==3.0.0
